+++ app/Models/Card.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
@@ -23,6 +24,21 @@ class Card extends Model
         return 'uuid';
     }
 
+    public function getSearchTextAttribute(): string
+    {
+        return collect([
+            $this->title,
+            $this->description,
+            ...$this->links->pluck('title')->toArray(),
+            ...$this->links->pluck('subtitle')->toArray(),
+        ])
+            ->filter()
+            ->map(fn($text) => mb_strtolower(trim($text)))
+            ->implode(' ');
+    }
+
+
+
     protected $fillable = [
         'uuid',
         'category_id',
@@ -38,6 +54,11 @@ class Card extends Model
         'is_active' => 'boolean',
         'expired_at' => 'datetime',
     ];
+
+    public function categories(): BelongsToMany
+    {
+        return $this->belongsToMany(Category::class);
+    }
 
     public function category(): BelongsTo
     {