+++ database/seeders/DatabaseSeeder.php
@@ -15,6 +15,12 @@ class DatabaseSeeder extends Seeder
      */
     public function run(): void
     {
-        $this->call(DemoSeeder::class);
+        $this->call([
+            PortalSettingSeeder::class,
+            DemoSeeder::class,
+        ]);
+            
+
+        
     }
 }