+++ resources/views/components/homepage/card.blade.php
@@ -1,17 +1,22 @@
 @props([
-    'card',
+'card',
 ])
 
 @if($card->links->count() === 1)
 
-    <x-homepage.single-card
-        :link="$card->links->first()" />
+<x-homepage.single-card
+    :link="$card->links->first()" />
 
 @else
 
-<section class="portal-group">
+<section
+    class="portal-group"
+    data-card
+    data-type="group"
+    data-category="{{ $card->categories->pluck('slug')->implode(',') }}"
+    data-search="{{ $card->search_text }}">
 
-    <div class="portal-group-header parent-toggle">
+    <div class="portal-group-header" data-accordion>
 
         <div class="app-left">
 
@@ -39,16 +44,16 @@
 
         </div>
 
-        <i class="bi bi-chevron-down portal-arrow"></i>
+        <i class="bi bi-chevron-down portal-arrow" data-arrow></i>
 
     </div>
 
-    <div class="portal-group-body d-none">
+    <div class="portal-group-body" data-body>
 
         @foreach($card->links as $link)
 
-            <x-homepage.child-card
-                :link="$link"/>
+        <x-homepage.child-card
+            :link="$link" />
 
         @endforeach